Create FlagQuizGame.java
This is the code for the Flag Quiz Game Application. Its an android application of a flag quiz guessing game. When the app starts the screen appears with the first flag and three possible answer buttons underneath it. You click one of the buttons with a country name on it and if you are correct it loads another flag and its possible answers. if you get the answer wrong it disables the button you chose and lets you pick from the remaining two. if does this until you choose the correct answer. when a incorrect answer is selected it displays "Incorrect!" in red at the bottom of the screen and there is a flag animation that makes the flag shake when there is an incorrect guess. When a correct answer is picked the correct country named displays at the bottom of the screen in green and then the screen loads to the next question with a new flag and 3 possible answers.At the end of the game a dialog box pops up with your statistics for the game  of how many guesses it took you to complete the game and the percentage you got. Also from this dialog box there is a button that lets you reset the quiz.